Refactored entire tf package as _tf
