+++ src/_tf/model_components/ebgan.py
