+++ src/_tf/model_components/mnist.py
@@ -32,7 +32,7 @@ def batch_norm(x, training, epsilon=1e-5, momentum=0.9):
 
 def dense(x, out_units):
     kernel = tf.random_normal_initializer(mean=0.0, stddev=0.3)
-    #     return tf.layers.dense(x, out_units, activation=None)
+    #     return _tf.layers.dense(x, out_units, activation=None)
     return layers.fully_connected(x, out_units, activation_fn=None)
