+++ src/_tf/models_tf/base_infergan.py
