+++ src/_tf/models_tf/dcgan.py
@@ -3,9 +3,8 @@ import logging
 import numpy as np
 import tensorflow as tf
 
-from tf.model_components import encoder, decoder, disc
-from tf.model_components import losses
-from tf.models_tf import BaseModel
+from _tf.model_components.mnist import encoder, decoder
+from _tf.models_tf import BaseModel
 from exp_context import ExperimentContext
 
 H = ExperimentContext.Hyperparams
@@ -32,9 +31,9 @@ class Model(BaseModel):
 
         logger.info('%s: Model Definition complete' % repr(self))
 
-        # logger.info('%s: Model Params:' % repr(self))
-        # for param in tf.trainable_variables(self.model_scope):
-        #     logger.info(param)
+        logger.info('%s: Model Params:' % repr(self))
+        for param in tf.trainable_variables():
+            logger.info(param)
 
         self.__is_model_built = True
 
@@ -49,7 +48,7 @@ class Model(BaseModel):
             self.add_param_saver(network_name, param_list)
 
     def _define_placeholders(self):
-        self.ph_X = tf.placeholder(tf.float32, [None, H.input_size])
+        self.ph_X = tf.placeholder(tf.float32, [None, H.input_height, H.input_width, H.input_channel])
         self.ph_Z = tf.placeholder(tf.float32, [None, H.z_size])
         self.ph_plot_img = tf.placeholder(tf.float32, [None, None, None, 4])
 
@@ -57,30 +56,55 @@ class Model(BaseModel):
         with tf.variable_scope(self.model_scope, reuse=tf.AUTO_REUSE):
             # X - Z - X Iteration
             self.x_real = self.ph_X
-            self.z_real = encoder(self.x_real)
+            self.out_real, self.logits_real, self.entropy_logits_real, self.z_real = encoder(self.x_real)  # ask
             self.x_recon = decoder(self.z_real)
 
             # Z - X - Z Iteration
             self.z_rand = self.ph_Z
             self.x_fake = decoder(self.z_rand)
-            self.z_recon = encoder(self.x_fake)
+            self.out_fake, self.logits_fake, self.entropy_logits_fake, self.z_recon = encoder(self.x_fake)  # ask
 
             # Disc Iteration
-            self.logits_real = disc(self.x_real)
-            self.logits_fake = disc(self.x_fake)
+            # self.out_real,self.logits_real,_ = disc(self.x_real)
+            # self.out_fake,self.logits_fake,_ = disc(self.x_fake)
 
-    def _define_losses(self):
+    def _define_losses(self):  # ask
+        smooth = 0.95
         batch_size = tf.shape(self.ph_X)[0]
         self.x_recon_loss = tf.reduce_mean((self.x_real - self.x_recon) ** 2)
+        print('x_real', self.x_real.shape)
+        print('x_fake', self.x_recon.shape)
+        print('z_real', self.z_rand.shape)
+        print('z_fake', self.z_recon.shape)
+
         self.z_recon_loss = tf.reduce_mean((self.z_rand - self.z_recon) ** 2)
         # [B, 2]
+        logit_batch_size = tf.shape(self.entropy_logits_real)[0]
+        real_entropy_labels = tf.ones([logit_batch_size, 1])
+        fake_entropy_labels = tf.zeros([logit_batch_size, 1])
+
         real_labels = tf.ones([batch_size, 1])
         fake_labels = tf.zeros([batch_size, 1])
 
-        self.disc_loss_real = losses.sigmoid_cross_entropy_loss(real_labels, self.logits_real)
-        self.disc_loss_fake = losses.sigmoid_cross_entropy_loss(fake_labels, self.logits_fake)
+        self.disc_batch_loss_real = tf.reduce_mean(
+            tf.nn.sigmoid_cross_entropy_with_logits(labels=real_entropy_labels, logits=self.entropy_logits_real))
+        self.disc_batch_loss_fake = tf.reduce_mean(
+            tf.nn.sigmoid_cross_entropy_with_logits(labels=fake_entropy_labels, logits=self.entropy_logits_fake))
 
-        self.gen_loss = losses.sigmoid_cross_entropy_loss(real_labels, self.logits_fake)
+        self.gen_batch_loss = tf.reduce_mean(
+            tf.nn.sigmoid_cross_entropy_with_logits(labels=real_entropy_labels, logits=self.entropy_logits_fake))
+
+        self.disc_sample_loss_real = tf.reduce_mean(
+            tf.nn.sigmoid_cross_entropy_with_logits(labels=real_labels, logits=self.logits_real))
+        self.disc_sample_loss_fake = tf.reduce_mean(
+            tf.nn.sigmoid_cross_entropy_with_logits(labels=fake_labels, logits=self.logits_fake))
+
+        self.gen_sample_loss = tf.reduce_mean(
+            tf.nn.sigmoid_cross_entropy_with_logits(labels=real_labels, logits=self.logits_fake))
+
+        self.disc_loss_real = (self.disc_batch_loss_real + self.disc_sample_loss_real) / 2.0
+        self.disc_loss_fake = (self.disc_batch_loss_fake + self.disc_sample_loss_fake) / 2.0
+        self.gen_loss = (self.gen_batch_loss + self.gen_sample_loss) / 2.0
 
         self.encoder_loss = self.x_recon_loss + self.z_recon_loss
         self.decoder_loss = self.encoder_loss + 0 * self.gen_loss
@@ -99,7 +123,7 @@ class Model(BaseModel):
     def _define_scopes(self):
         self.encoder_scope = self.model_scope + '/encoder'
         self.decoder_scope = self.model_scope + '/decoder'
-        self.disc_scope = self.model_scope + '/disc'
+        self.disc_scope = self.model_scope + '/encoder'
 
     def _define_summaries(self):
         summaries_list = [
@@ -117,8 +141,8 @@ class Model(BaseModel):
             tf.summary.histogram('z_real', self.z_real),
             tf.summary.histogram('z_recon', self.z_recon),
 
-            # tf.summary.histogram('fake_preds', self.disc_real_preds),
-            # tf.summary.histogram('real_preds', self.disc_fake_preds),
+            # _tf.summary.histogram('fake_preds', self.disc_real_preds),
+            # _tf.summary.histogram('real_preds', self.disc_fake_preds),
         ]
 
         self.summaries = tf.summary.merge(summaries_list)
@@ -128,21 +152,20 @@ class Model(BaseModel):
         self.add_param_group('encoder', tf.global_variables(self.model_scope + '/encoder'))
         self.add_param_group('decoder', tf.global_variables(self.model_scope + '/decoder'))
         self.add_param_group('autoencoder', self.param_groups['encoder'] + self.param_groups['decoder'])
-        self.add_param_group('disc', tf.global_variables(self.model_scope + '/disc'))
         self.add_param_group('all', tf.global_variables(self.model_scope))
 
     def _define_operations(self):
-        with tf.variable_scope('', reuse=tf.AUTO_REUSE):
-            autoencoder_trainable_params = tf.trainable_variables(self.encoder_scope) + tf.trainable_variables(self.decoder_scope)
-            opt = tf.train.RMSPropOptimizer(H.lr_autoencoder)
+        with tf.variable_scope(self.model_scope, reuse=tf.AUTO_REUSE):
+            autoencoder_trainable_params = tf.global_variables(self.encoder_scope) + tf.global_variables(self.decoder_scope)
+            opt = tf.train.AdamOptimizer(H.lr_autoencoder, beta1=H.beta1, beta2=H.beta2)
             self.autoencoder_train_op = opt.minimize(self.encoder_loss, var_list=autoencoder_trainable_params)
 
-            decoder_params = tf.trainable_variables(self.decoder_scope)
-            opt = tf.train.RMSPropOptimizer(H.lr_autoencoder)
+            decoder_params = tf.global_variables(self.decoder_scope)
+            opt = tf.train.AdamOptimizer(H.lr_autoencoder, beta1=H.beta1, beta2=H.beta2)
             self.adv_gen_train_op = opt.minimize(self.gen_loss, var_list=decoder_params)
 
-            disc_params = tf.trainable_variables(self.disc_scope)
-            opt = tf.train.RMSPropOptimizer(H.lr_autoencoder)
+            disc_params = tf.global_variables(self.disc_scope)
+            opt = tf.train.AdamOptimizer(H.lr_autoencoder, beta1=H.beta1, beta2=H.beta2)
             self.adv_disc_train_op = opt.minimize(self.disc_loss, var_list=disc_params)
 
     @property
@@ -212,10 +235,10 @@ class Model(BaseModel):
     def discriminate(self, x_batch, split=True):
         preds = self.session.run(self.disc_real_preds, {
             self.ph_X: x_batch
-        })[:, 0]
+        })
         if split:
-            x_batch_real = x_batch[np.where(preds == 1)]
-            x_batch_fake = x_batch[np.where(preds == 0)]
+            x_batch_real = x_batch[np.where(preds == 0)]
+            x_batch_fake = x_batch[np.where(preds == 1)]
             return preds, x_batch_real, x_batch_fake
         else:
             return preds
